Add files via upload
This Lab exercises are the Azure vertiual machines creating images and creatin vertual networking.  for the providing the student for free get azure vertual machines and more .it help to get first expirence to undergadues students or first cloud fundamental learning students, get azuru related expirence. as group project , we were done things are available as pdf  and as personal as Lab